Temp. disable remote building from workbench
This permits us to cut a stable version with most of the remote build
support in it without announcing the capability.  This in turn enables
merging changes in the runtimes and main and deploying them in nimgcp
without the need to do it all over again in a second pass.  At that
point we can document and announce.