Support for remote build (remains hidden initially while backend support is completed) (#17)
* Add command authoring guide

* WIP for remote building

* WIP for remote building

* WIP for remote building

* Bring remote-build branch up to date

It had gotten quite stale and need reworking to accommodate intervening
changes.

* WIP for remote building

* Small adjustments for remote build.

Remote build now successfully uploads to client data bucket.  The
builder service does not know to look for it there (yet).

* WIP for remote building

* Add slice reader, support local testing

* WIP for remote building

* Remote builder working on nodejs runtimes

* WIP for remote build, interim commit

* Get remote builds working for actionloop runtimes

* Hide remote build flag to support possible merge

The functionality is not ready for support and won't be documented for
a while but may be present in the next stable version to facilitate
roll-out.

* Periodic ticks while waiting for remote build

* Use simpler name for slice cache directory

* Create utility to write the slice result

This is needed to separate the output from debugging messages, which
will need to go to stdout in order to appear in activation logs.  It
is not yet clear where the slice result really should go and so this is
not yet being called.

* "Failures:" is a header, not an error itself

* Make build failures non-terminal where possible

* Change handling of deployer annotation

Supports propagating a deployer annotation along with a slice so that
it comes out right in a remote build.

* Improve display of project paths in remote build

* Fixes to get correct behavior in workbench

* Add missing quote mark

* Further fixes for workbench remote build

* Increase memory buffer size for slice uploads

* Temp. disable remote building from workbench

This permits us to cut a stable version with most of the remote build
support in it without announcing the capability.  This in turn enables
merging changes in the runtimes and main and deploying them in nimgcp
without the need to do it all over again in a second pass.  At that
point we can document and announce.

* Correct commander-cli githash