Attempt to fix automated test workflow
It was broken by the the change to no longer assume apigcp as a default
API host for login. In retrospect, this should have been handled as a
breaking change or the legacy behavior preserved a bit longer.  But,
moving ahead instead of going back.  Adding an -apihost flag should do
it.