Pick up required deployer change and incr. version
This new version will fix bugs in both deployer and CLI and add the
check when creating a new project that a runtime actually exists for it
in the backend.   the check only occurs if there is an API host already
set at project creation time, which is not a requirement but is usually
true.