Project create should check that a runtime exists for a project before creating it (#219)
* Fix two bugs in 'nim info --runtimes' command

1.  Logic was confused and output sometimes did not print due to a
missing 'await'.  2.  Table formatting was off if shortest runtime name
was shorter than the word "language" in the header.

* Project create logic will check runtime validity

Now refuses to create a project that cannot be deployed.  No longer
displays an explicit hardcoded enumeration of languages.

* 2.3.8

* Pick up required deployer change and incr. version

This new version will fix bugs in both deployer and CLI and add the
check when creating a new project that a runtime actually exists for it
in the backend.   the check only occurs if there is an API host already
set at project creation time, which is not a requirement but is usually
true.