Revert "3.0.3"
This reverts commit 060bc50b654d5ae865d09769e2e5a9739af72086.
We need to abort the incorporation of deployer 4.0.4 and fix
the deployer dependency to be locked down.  We can then
reuse the 3.0.3 version because it was never released to
npm.  That deployer change is breaking unless the builder
actions are changed.  It should not be incorporated until
the builder changes are in place in the target deployments.