Use 'functions' not 'actions' in project.yml
This is a breaking change for tools that parse the output of 'project
get-metadata'.  On the input side, it is backward compatible (both
'functions' and 'actions' are accepted).  For human consumption, the
difference may be noticed but should not be too disruptive.  Major
version 4.0.0 declared because of the potential breakage.